+++ ttnn/core/services/d2h_socket_service.cpp
@@ -59,6 +59,83 @@ distributed::MeshComposerConfig derive_composer_config(const distributed::MeshMa
     return distributed::MeshComposerConfig{.dims = dims, .mesh_shape_override = mapper_config.mesh_shape_override};
 }
 
+// ---------------------------------------------------------------------------
+// D2H split-kernel chunk plan (reader/writer pipeline).
+//
+// Like the shared ChunkPlan but adds slot_count: the data-CB depth (in full
+// socket-page slots) that lets the reader stage DRAM pages ahead while the
+// writer drains earlier ones into the host-pinned socket FIFO. Mirrors the
+// H2DChunkPlan approach: size the socket page to a few NOC bursts (capped by
+// max_socket_page_size_bytes), then fill the remaining service-core L1 with
+// slots above a double-buffering floor.
+// ---------------------------------------------------------------------------
+
+constexpr uint32_t kD2HNocBurstBytes = 16u * 1024;
+constexpr uint32_t kD2HTargetReadBursts = 8;  // default socket-page target ~= 8 bursts (128 KB)
+constexpr uint32_t kD2HSlotCap = 64;          // upper bound on data-CB slots
+constexpr uint32_t kD2HMinDataSlots = 2;      // double-buffering floor (reader/writer overlap)
+
+// The data CB (program allocator, bottom-up) and the service-core scratch
+// (ServiceCoreManager, top-down) share the unreserved L1 with no cross-allocator
+// overflow check. This reserve holds back headroom for the socket config buffer
+// and post-plan scratch words (termination, worker-sync counters, metadata
+// staging) plus a safety pad. Generous on purpose (>> their sum).
+constexpr uint64_t kD2HServiceScratchReserveBytes = 16u * 1024;
+
+struct D2HChunkPlan {
+    uint32_t socket_page_size;  // bytes per socket page (== pages_per_chunk * tensor_page_size)
+    uint32_t num_socket_pages;  // socket pages per full transfer (== tensor_num_pages / pages_per_chunk)
+    uint32_t pages_per_chunk;   // tensor pages drained per socket page
+    uint32_t slot_count;        // full-page data-CB slots backing the reader/writer pipeline
+};
+
+D2HChunkPlan derive_d2h_chunk_plan(
+    uint32_t tensor_page_size, uint32_t tensor_num_pages, uint64_t usable_cb_l1_bytes, uint64_t page_budget_hint) {
+    TT_FATAL(tensor_page_size > 0, "D2HStreamService: tensor page size must be > 0");
+    TT_FATAL(tensor_num_pages > 0, "D2HStreamService: tensor must have at least one page");
+    TT_FATAL(
+        tensor_page_size <= usable_cb_l1_bytes,
+        "D2HStreamService: tensor page {} B exceeds service-core CB L1 budget {} B; "
+        "use a layout with smaller pages",
+        tensor_page_size,
+        usable_cb_l1_bytes);
+    TT_FATAL(
+        page_budget_hint == 0 || page_budget_hint >= tensor_page_size,
+        "D2HStreamService: max_socket_page_size_bytes ({} B) must be >= the tensor page size ({} B); "
+        "the socket page can't be smaller than one tensor page (pass 0 to auto-size)",
+        page_budget_hint,
+        tensor_page_size);
+
+    const uint64_t burst_target = static_cast<uint64_t>(kD2HTargetReadBursts) * kD2HNocBurstBytes;
+    const uint64_t requested = page_budget_hint > 0 ? page_budget_hint : burst_target;
+    const uint64_t page_budget = std::min<uint64_t>(requested, usable_cb_l1_bytes / kD2HMinDataSlots);
+
+    uint32_t pages_per_chunk = std::max<uint32_t>(1, static_cast<uint32_t>(page_budget / tensor_page_size));
+    pages_per_chunk = std::min(pages_per_chunk, tensor_num_pages);
+    while (pages_per_chunk > 1 && (tensor_num_pages % pages_per_chunk) != 0) {
+        --pages_per_chunk;
+    }
+    const uint32_t socket_page_size = pages_per_chunk * tensor_page_size;
+
+    const uint32_t slots_l1_max = static_cast<uint32_t>(usable_cb_l1_bytes / socket_page_size);
+    const uint32_t slot_count = std::min(slots_l1_max, kD2HSlotCap);
+    if (slot_count < kD2HMinDataSlots) {
+        log_warning(
+            tt::LogOp,
+            "D2HStreamService: tensor page {} B leaves only {} CB slot(s) in {} B of L1; "
+            "reader/writer overlap disabled (use a smaller per-shard page to double-buffer)",
+            tensor_page_size,
+            slot_count,
+            usable_cb_l1_bytes);
+    }
+    return D2HChunkPlan{
+        .socket_page_size = socket_page_size,
+        .num_socket_pages = tensor_num_pages / pages_per_chunk,
+        .pages_per_chunk = pages_per_chunk,
+        .slot_count = slot_count,
+    };
+}
+
 struct D2HWorkerSyncArgs {
     bool enabled = false;
     uint32_t transfer_done_sem_addr = 0;
